+++ src/index.js
@@ -1,8 +1,5 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Login from './page/auth/login';
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "./App";
 
-
-ReactDOM.render(<Login />,document.getElementById('root'));
-
-
+ReactDOM.render(<App />, document.getElementById("root"));
