--- src/page/auth/login.js
@@ -1,24 +1,0 @@
-import React from 'react';
-import './common.css';
-import { Button, ButtonGroup } from "@chakra-ui/react";
-
-function Login() {
-    return (
-        <div className="body">
-            <div className = "container">
-                <div className="app">
-                    <div className="logo"></div>
-                    <h1 className="LoginHeader">
-                        개인아이디로 로그인하세요
-                    </h1>
-                    <ButtonGroup className="button-Group">
-                        <Button className="blue">스트리머 로그인</Button>
-                        <Button className="white">시청자 로그인</Button>
-                    </ButtonGroup>
-                </div>
-            </div>
-        </div>
-    );
-}
-
-export default Login;
